requirement.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.4
mysql-connector==2.2.9
mysql-connector-python==8.0.23
mysql-connector-python-rf==2.2.2
numpy==1.20.1
Pillow==8.1.0
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
